refactor: migrate keychain store
