+++ .gitignore
@@ -13,3 +13,4 @@ coverage
 architecture/*.svg
 *.code-workspace
 tmp
+**/tsup.config.native.bundled*
