+++ apps/mobile/src/store/keychains/keychains.read.ts
@@ -1,5 +1,6 @@
+import { keychainAdapter } from '@leather.io/state/keychains';
+
 import { RootState } from '..';
-import { keychainAdapter } from './keychains.write';
 
 export const keychainSelectors = keychainAdapter.getSelectors(
   (state: RootState) => state.keychains
